pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "assignment"
version = "0.0.1"
dependencies = [
"warp-lang==1.9.1",
"torch",
"numpy",
"panda3d==1.10.15",
"pytest",
"polyscope",
"libigl",
"usd-core",
"sympy",
"h5py",
]


[tool.setuptools]
packages = ["assignment", "utils"]
